Get us to 0 warnings 0 errors on all platforms (#1089)
* Fix compiler warnings on Windows

Also, since downstream I upgraded catch, it seems more picky about what can go inside of its macros.  In particular C4Slice comparison using == doesn't work, and needs to be inside another set of parenthesis.  In particular the following sets of warnings were dealt with:

1. Lots and lots of warnings about possible loss of data due to size truncation (added narrow_cast as a way to both do this without the warning, and get some bounds checking in debug builds to make sure we aren't passing in bad data)
2. Unsafe use of bool (int i = false; i <= true)
3. Function doesn't always return a false (switch with no default)
4. Unused variable (inside of catch blocks and a few places where comments didn't comment the whole block)
5. 32-bit shift expanded to 64-bit variable
6. (Windows specific) Having objects defined in the compilation unit that are marked __declspec(dllimport)

+++ Crypto/PublicKey+Windows.cc
@@ -31,6 +31,7 @@
 #include "mbedtls/md5.h"
 #include "TempArray.hh"
 #include "Defer.hh"
+#include "NumConversion.hh"
 #include <Windows.h>
 #include <ncrypt.h>
 #include <wincrypt.h>
@@ -80,7 +81,7 @@ namespace litecore::crypto {
         PCCERT_CONTEXT resultCert = CertCreateCertificateContext(
             X509_ASN_ENCODING|PKCS_7_ASN_ENCODING,
             (const BYTE *)cert->data().buf,
-            cert->data().size
+            narrow_cast<DWORD>(cert->data().size)
         );
 
         if(!resultCert) {
@@ -334,7 +335,7 @@ namespace litecore::crypto {
                 _keyLength,
                 nullptr,
                 PBYTE(output),
-                output_max_len,
+                narrow_cast<DWORD>(output_max_len),
                 reinterpret_cast<DWORD*>(output_len),
                 NCRYPT_PAD_PKCS1_FLAG
             );
@@ -387,7 +388,7 @@ namespace litecore::crypto {
                 _keyPair,
                 &padding,
                 PBYTE(inputData.buf),
-                inputData.size,
+                narrow_cast<DWORD>(inputData.size),
                 PBYTE(outSignature),
                 _keyLength,
                 &dummy,
@@ -539,7 +540,7 @@ namespace litecore::crypto {
 
             if(cert == this) {
                 CRYPT_DATA_BLOB idBlob {
-                    persistentID.size() + 1,
+                    narrow_cast<DWORD>(persistentID.size() + 1),
                     PBYTE(persistentID.c_str())
                 };
