+++ LiteCore/Storage/UnicodeCollator_winapi.cc
@@ -23,6 +23,7 @@
 #include "SQLiteCpp/SQLiteCpp.h"
 #include "StringUtil.hh"
 #include "TempArray.hh"
+#include "NumConversion.hh"
 #include <sqlite3.h>
 #include <algorithm>
 
@@ -61,7 +62,7 @@ namespace litecore {
             else {
                 string tmp((const char*)localeSlice.buf, localeSlice.size);
                 replace(tmp, '_', '-');
-                MultiByteToWideChar(CP_UTF8, 0, tmp.c_str(), tmp.size(), localeName, LOCALE_NAME_MAX_LENGTH);
+                MultiByteToWideChar(CP_UTF8, 0, tmp.c_str(), narrow_cast<DWORD>(tmp.size()), localeName, LOCALE_NAME_MAX_LENGTH);
                 if (LocaleNameToLCID(localeName, 0) == 0) {
                     Warn("Unknown locale name '%.*s', using default", SPLAT(coll.localeName));
                     LCID lcid = MAKELCID(MAKELANGID(LANG_ENGLISH, SUBLANG_ENGLISH_US), SORT_DEFAULT);
